move other applicable external libraries to separate projects
