+++ xlive/3rdparty/tinyxml/tinyxml/tinyxml2.cpp
